Rename js files to ts
Update .gitignore to ensure dist folder is not commited
