+++ services/user/src/index.ts
@@ -1,8 +1,8 @@
 import express from "express";
 import cors from "cors";
 
-import userRoutes from "./routes/user-routes.js";
-import authRoutes from "./routes/auth-routes.js";
+import userRoutes from "./routes/user-routes";
+import authRoutes from "./routes/auth-routes";
 
 const app = express();
