+++ services/user/src/routes/auth-routes.ts
@@ -1,7 +1,7 @@
 import express from "express";
 
-import { handleLogin, handleVerifyToken } from "../controller/auth-controller.js";
-import { verifyAccessToken } from "../middleware/basic-access-control.js";
+import { handleLogin, handleVerifyToken } from "../controller/auth-controller";
+import { verifyAccessToken } from "../middleware/basic-access-control";
 
 const router = express.Router();
